fix: Refactor `PHPDoc` comments for consistency and clarity.
