+++ tests/type/ActiveQueryDynamicReturnTypeExtensionTest.php
@@ -37,7 +37,7 @@ final class ActiveQueryDynamicReturnTypeExtensionTest extends TypeInferenceTestC
     {
         $directory = dirname(__DIR__);
 
-        yield from self::gatherAssertTypes("{$directory}/fixture/data/types/ActiveQueryDynamicMethodReturnType.php");
+        yield from self::gatherAssertTypes("{$directory}/fixture/data/type/ActiveQueryDynamicMethodReturnType.php");
     }
 
     public static function getAdditionalConfigFiles(): array
